// src/api/Learnup/models/AudioBookResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type AudioBookResponse = {
    id: number;
    title: string;
    description: string | null;
    author: string | null;
    level: string | null;
    year: string | null;
    wordCount: string | null;
    source: string | null;
    coverId: string | null;
    isVoiced: boolean;
};